refactor(cli): split repl.ts monolith into focused modules
Extract createConfirmFn → confirm.ts, runAgentTurn/createSpinner →
runner.ts, and the plan flow → plan.ts, leaving repl.ts as a thin
orchestration loop. Each module owns one concern, making them
independently testable and readable.

Closes #52
