+++ src/cli/index.ts
@@ -7,7 +7,8 @@ import { loadConfig, saveConfig, AGENT_DIR } from "../state/config.js";
 import { Session } from "../state/session.js";
 import { loadSystemInstruction } from "../core/prompt.js";
 import { resolveApiKey } from "./keys.js";
-import { runRepl, createConfirmFn } from "./repl.js";
+import { runRepl } from "./repl.js";
+import { createConfirmFn } from "./confirm.js";
 import { printError, printInfo } from "./renderer.js";
 import type { ObservabilityEvent } from "../core/observability.js";
 import { createSandboxRunner } from "../tools/exec/sandbox/index.js";
